Cleanup default branch check
